Change around how prefixes are configured.
It makes more sense to have a prefix be the primary unit, and have
remote repositories be associated with a prefix.  While here move a lot
more of the prefix setup into config rather than doing it in e.g.
update, which will help future commands.

Update the example configuration to be more realistic.